update
inumがうまくうごかない。fatになるのでserviceにまかせた。
