--- app/enum/company_yahoo_scraping_nums.rb
@@ -1,10 +1,0 @@
-class CompanyYahooScapingNums < Inum::Base
-  define :ADMIN, 777
-  define :GENERAL, 0
-  # definition
-  define :REG_EXP_PERIOD_ENDING_TITLE, 30
-  define :REG_EXP_PERIOD_ENDING_PQ_ONE, 32
-  define :REG_EXP_PERIOD_ENDING_PQ_TWO, 34
-  define :REG_EXP_PERIOD_ENDING_PQ_THREE, 36
-  define :REG_EXP_PERIOD_ENDING_PQ_FOUR, 38
-end
