+++ app/enums/user_cans.rb
